backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
slither-analyzer==0.9.3
transformers==4.35.0
torch==2.6.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
py-solc-x==1.1.1
web3==6.11.0
gunicorn==21.2.0 